# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CampusHub API
  version: 0.1.0
  description: API contract for the CampusHub mini-program, admin dashboard, and backend.
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /users/{userId}/followers:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [User]
      summary: User followers
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paged followers
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserPage"
  /users/{userId}/following:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [User]
      summary: Users followed by user
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paged followed users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserPage"
...
  /posts/{postId}/like:
    parameters:
      - $ref: "#/components/parameters/PostId"
    post:
      tags: [Interaction]
      summary: Like a post idempotently
      responses:
        "200":
          description: Current like state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InteractionStateResponse"
    delete:
      tags: [Interaction]
      summary: Unlike a post idempotently
      responses:
        "200":
          description: Current like state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InteractionStateResponse"
  /users/{userId}/follow:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Interaction]
      summary: Follow user idempotently
      responses:
        "200":
          $ref: "#/components/responses/OkNull"
    delete:
      tags: [Interaction]
      summary: Unfollow user idempotently
      responses:
        "200":
          $ref: "#/components/responses/OkNull"
  /messages:
    get:
      tags: [Message]
      summary: Current user's inbox
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paged messages
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/MessagePage"
...
  /messages/{messageId}/read:
    put:
      tags: [Message]
      summary: Mark one message as read
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/OkNull"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
    PostId:
      name: postId
      in: path
      required: true
      schema:
        type: string
  responses:
    OkNull:
      description: Operation succeeded
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
              - type: object
                properties:
                  data:
                    nullable: true
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message, data, traceId]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: ok
        data: {}
        traceId:
          type: string
...
    User:
      type: object
      required: [userId, nickname, avatar, bio, role, status]
      properties:
        userId:
          type: string
        nickname:
          type: string
        avatar:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        role:
          $ref: "#/components/schemas/UserRole"
        status:
          $ref: "#/components/schemas/UserStatus"
...
    Message:
      type: object
      required: [messageId, fromUser, type, targetType, targetId, content, isRead, createTime]
      properties:
        messageId:
          type: string
        fromUser:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/User"
        type:
          type: integer
        targetType:
          $ref: "#/components/schemas/TargetType"
        targetId:
          type: string
        content:
          type: string
          nullable: true
        isRead:
          type: boolean
        createTime:
          type: string
          format: date-time
...
    PageMeta:
      type: object
      required: [total, page, size, hasMore]
      properties:
        total:
          type: integer
        page:
          type: integer
        size:
          type: integer
        hasMore:
          type: boolean
...
    UserPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [list]
          properties:
            list:
              type: array
              items:
                $ref: "#/components/schemas/User"
...
    MessagePage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [list]
          properties:
            list:
              type: array
              items:
                $ref: "#/components/schemas/Message"
...
    InteractionStateResponse:
      allOf:
        - $ref: "#/components/schemas/ApiResponse"
        - type: object
          properties:
            data:
              type: object
              required: [liked, likeCount]
              properties:
                liked:
                  type: boolean
                likeCount:
                  type: integer
    UserRole:
      type: integer
      enum: [0, 1, 2]
      description: 0 student, 1 moderator, 2 admin
    UserStatus:
      type: integer
      enum: [0, 1]
...
    TargetType:
      type: integer
      enum: [1, 2, 3]